Create Quiz_Master
Welcome to "Quiz Master" - an interactive quiz game that will test your knowledge and challenge your trivia skills! Are you ready to put your thinking cap on and dive into an exciting world of questions? This project is designed to provide you with an engaging quiz experience, where you'll have the opportunity to answer a series of 10 questions on various topics. Each correct answer will earn you a point, and there's no need to worry about negative marking.

Once you complete the quiz, "Quiz Master" will display your final score and deliver a personalized message based on your performance. If you manage to score 6 points or above, you'll be treated to a cheerful celebration, acknowledging your achievement as a true quiz whiz. On the other hand, if you fall short of 6 points, don't be discouraged! "Quiz Master" will offer words of encouragement, reminding you to keep learning and giving it another shot.

So, are you up for the challenge? Let's embark on this quiz adventure and put your knowledge to the test. Get ready to have fun, learn new things, and see if you have what it takes to become the ultimate "Quiz Master"!